requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
xrpl-py>=4.1.0
pydantic>=2.0.0
